refactor: extract FileLoader from Main.kt and add computeMaxDistanceFromStart function
